+++ semcore/base-trigger/src/BaseTrigger.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import createComponent, { Component, Root, sstyled } from '@semcore/core';
 import { Box } from '@semcore/flex-box';
-import { neighborLocationEnhance } from '@semcore/neighbor-location';
+import { NEIGHBOR_LOCATION_AUTO_DETECT } from '@semcore/neighbor-location';
 import keyboardFocusEnhance from '@semcore/utils/lib/enhances/keyboardFocusEnhance';
 import addonTextChildren from '@semcore/utils/lib/addonTextChildren';
 import logger from '@semcore/utils/lib/logger';
@@ -10,11 +10,12 @@ import style from './style/base-trigger.shadow.css';
 
 class RootBaseTrigger extends Component {
   static displayName = 'BaseTrigger';
+  static enhance = [keyboardFocusEnhance()];
+  static [NEIGHBOR_LOCATION_AUTO_DETECT] = true;
   static style = style;
   static defaultProps = {
     size: 'm',
   };
-  static enhance = [keyboardFocusEnhance(), neighborLocationEnhance()];
 
   getTextProps() {
     const { placeholder, empty } = this.asProps;