+++ website/docs/utils/neighbor-location/examples/neighbor-location-with-custom.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import NeighborLocation, { NEIGHBOR_LOCATION_AUTO_DETECT } from '@semcore/neighbor-location';
+
+function CustomComponent({ neighborLocation }) {
+  return neighborLocation;
+}
+
+CustomComponent[NEIGHBOR_LOCATION_AUTO_DETECT] = true;
+
+const Demo = () => {
+  return (
+    <NeighborLocation>
+      <CustomComponent />
+      <CustomComponent />
+      <CustomComponent />
+    </NeighborLocation>
+  );
+};
+
+export default Demo;